Wydzielona PCL
