+++ WalkaChomika.Common/Models/ArmiaChomików.cs
@@ -6,11 +6,10 @@ using System.Threading.Tasks;
 
 namespace WalkaChomika.Models
 {
-
     /// <summary>
     /// Definicja nowego zawodnika - Armii Chomików
     /// </summary>
-    class ArmiaChomików : Zwierzę
+    public class ArmiaChomików : Zwierzę
     {
         /// <summary>
         /// Wewnętrzna kolekcja, lista, zawierająca poszczególne Chomiki
@@ -30,8 +29,8 @@ namespace WalkaChomika.Models
                 {
                     h += c.HP;
                 }
-                
-                return h;                
+
+                return h;
             }
 
             set
@@ -92,7 +91,7 @@ namespace WalkaChomika.Models
         /// <returns>Zwraca, czy istnieje przynajmniej jeden chomik w armii.</returns>
         public override bool CzyŻyje()
         {
-            return (chomiki.Count > 0);              
+            return (chomiki.Count > 0);
         }
 
         /// <summary>
