+++ WalkaChomika.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 12.0.30723.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalkaChomika", "WalkaChomika\WalkaChomika.csproj", "{6DBC7483-A395-4721-89B1-C9320DBB808F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WalkaChomika.Common", "WalkaChomika.Common\WalkaChomika.Common.csproj", "{48C6773B-3868-426A-B1CA-61F139D410FC}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{6DBC7483-A395-4721-89B1-C9320DBB808F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6DBC7483-A395-4721-89B1-C9320DBB808F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6DBC7483-A395-4721-89B1-C9320DBB808F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{48C6773B-3868-426A-B1CA-61F139D410FC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{48C6773B-3868-426A-B1CA-61F139D410FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{48C6773B-3868-426A-B1CA-61F139D410FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{48C6773B-3868-426A-B1CA-61F139D410FC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
