+++ WalkaChomika/WalkaChomika.csproj
@@ -56,11 +56,6 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="Models\ArmiaChomików.cs" />
-    <Compile Include="Models\ChomikSzaman.cs" />
-    <Compile Include="Models\ILatający.cs" />
-    <Compile Include="Models\Jednorożec.cs" />
-    <Compile Include="Models\ZwierzęMagiczne.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -69,10 +64,7 @@
       <DependentUpon>App.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Models\Chomik.cs" />
-    <Compile Include="Models\Dinozaur.cs" />
     <Compile Include="TextBoxTraceListener.cs" />
-    <Compile Include="Models\Zwierzę.cs" />
     <Compile Include="MainWindow.xaml.cs">
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
@@ -104,6 +96,12 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\WalkaChomika.Common\WalkaChomika.Common.csproj">
+      <Project>{48c6773b-3868-426a-b1ca-61f139d410fc}</Project>
+      <Name>WalkaChomika.Common</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
